pyproject.toml
[project]
name = "gitlab-mcp"
version = "0.1.0"
description = "GitLab MCP server - AI-optimized interface to GitLab"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0,<3",
    "python-gitlab>=4.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "pyrefly>=0.52.0",
    "basedpyright>=1.38.0",
    "ruff>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
gitlab-mcp = "gitlab_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gitlab_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.basedpyright]
typeCheckingMode = "standard"
# Only suppress the noisiest warnings from missing library stubs
reportUnknownMemberType = "warning"
reportUnknownArgumentType = "warning"
reportUnknownVariableType = "warning"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
